Property Info code pushed

+++ src/main/java/com/razysave/service/serviceImpl/property/PropertyServiceImpl.java
@@ -1,16 +1,23 @@
 package com.razysave.service.serviceImpl.property;
 
+import com.razysave.dto.device.DeviceListDto;
+import com.razysave.dto.property.PropertyDto;
+import com.razysave.dto.unit.UnitListDto;
 import com.razysave.entity.property.Building;
 import com.razysave.entity.property.Property;
+import com.razysave.entity.property.Unit;
+import com.razysave.entity.tenant.Tenant;
 import com.razysave.exception.PropertyNotFoundException;
 import com.razysave.repository.property.PropertyRepository;
 import com.razysave.service.property.BuildingService;
 import com.razysave.service.property.PropertyService;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class PropertyServiceImpl implements PropertyService {
@@ -18,10 +25,19 @@ public class PropertyServiceImpl implements PropertyService {
     private PropertyRepository propertyRepository;
     @Autowired
     private BuildingService buildingService;
+    private ModelMapper modelMapper = new ModelMapper();
 
-    public List<Property> getProperties() {
+    public List<PropertyDto> getProperties() {
         List<Property> properties = propertyRepository.findAll();
-        return properties;
+        if(properties.isEmpty())
+        {
+            throw new PropertyNotFoundException("No Property Found");
+        }
+        else {
+            return properties.stream()
+                    .map(this::mapToDto)
+                    .collect(Collectors.toList());
+        }
     }
 
     public Property getPropertyById(Integer id) {
@@ -35,7 +51,8 @@ public class PropertyServiceImpl implements PropertyService {
 
 
     public Property addProperty(Property property) {
-        return propertyRepository.save(property);
+       property.setId(4);
+       return propertyRepository.save(property);
     }
 
     public Property updateProperty(Integer id, Property updatedproperty) {
@@ -88,13 +105,19 @@ public class PropertyServiceImpl implements PropertyService {
         if (propertyOptional.isPresent()) {
             Property property = propertyOptional.get();
             List<Building> buildings = property.getBuilding();
-            for (Building building : buildings) {
-                buildingService.deleteBuildingById(building.getId());
+            if(buildings!=null) {
+                for (Building building : buildings) {
+                    buildingService.deleteBuildingById(building.getId());
+                }
             }
             propertyRepository.deleteById(id);
         } else {
             throw new PropertyNotFoundException("Property Not found with id :" + id);
         }
+    }
+    private PropertyDto mapToDto(Property property) {
+        PropertyDto dto = modelMapper.map(property, PropertyDto.class);
+        return dto;
     }
 }
